parser_vit/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.1.2
django-autoslug==1.9.8
django-dash==0.6.1
django-nine==0.2.6
django-plotly-dash==2.1.3
django-tinymce==3.5.0
djangorestframework==3.14.0
dpd-components==0.1.0
easy-thumbnails==2.8.3
feedparser==6.0.10
Flask==2.2.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.1
mod-wsgi==4.9.4
numpy==1.23.5
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pif==0.8.2
Pillow==9.3.0
plotly==5.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.6.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tenacity==8.1.0
tqdm==4.64.1
transliterate==1.10.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
Unidecode==1.3.6
urllib3==1.26.12
vishap==0.1.5
webdriver-manager==3.8.5
Werkzeug==2.2.2
wsproto==1.2.0
